+++ src/test/java/de/unistuttgart/iste/meitrex/assignment_service/api/AuthorizationTest.java
@@ -1,6 +1,6 @@
 package de.unistuttgart.iste.meitrex.assignment_service.api;
 
-import de.unistuttgart.iste.meitrex.assignment_service.persistence.entity.AssignmentEntity;
+import de.unistuttgart.iste.meitrex.assignment_service.persistence.entity.assignment.AssignmentEntity;
 import de.unistuttgart.iste.meitrex.assignment_service.persistence.repository.AssignmentRepository;
 import de.unistuttgart.iste.meitrex.assignment_service.test_utils.TestUtils;
 import de.unistuttgart.iste.meitrex.common.testutil.AuthorizationAsserts;
